+++ ball.h
@@ -7,9 +7,9 @@ class Ball {
         float position_y;
         float velocity_x;
         float velocity_y;
-        int id;
+        float id;
         Ball();
-        Ball(float r , float m, float pos_x, float pos_y, float vel_x, float vel_y, int id);
+        Ball(float r , float m, float pos_x, float pos_y, float vel_x, float vel_y, float id);
 
         void update(float dt);
